pyproject.toml
[project]
name = "hydra-wandb"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gokamoda", email = "gokamoda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "rich>=13.9.4",
    "hydra-core>=1.3.2",
    "pytz>=2025.2",
    "tqdm>=4.67.1",
    "multiprocess>=0.70.18",
    "wandb>=0.21.3",
    "numpy>=2.3.2",
]


[tool.uv]
dev-dependencies = [
    "ipython>=8.29.0",
    "mypy>=1.11.2",
    "pre-commit>=3.8.0",
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]
conflicts = [
  [
    { extra = "default" },
    { extra = "cpu" },
    { extra = "cu118" },
    { extra = "cu126" },
  ],
]


[project.optional-dependencies]
default = [
  "torch==2.7.0",
]
cpu = [
  "torch==2.7.0",
]
cu118 = [
  "torch==2.7.0",
]
cu126 = [
  "torch==2.7.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu118", extra = "cu118" },
  { index = "pytorch-cu126", extra = "cu126" },
]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["src",]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "B905",  # Zip-without-explicit-strict
    "E501",  # Line too long, handled by black
    "F403",  # Unable to detect undefined names
    "N812",  # Lowercase imported as non-lowercase
    "N816",  # Global scope should not be mixedCase
    "N803",  # Argument name should be lowercase
    "N806",  # Variable in function should be lowercase
    "N999",  # Invalid module name
    "UP035",  # Checks for uses of deprecated imports based on the minimum supported Python version.
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR0915",  # Too many statements in function
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true  # Settings for Python 3.8 compatibility

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
